requirements.txt
numpy
pandas
scikit-learn
sentence-transformers==2.2.2
python-rake==1.4.4
torch==1.10.1
tqdm
